+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/controller/ItemSearchController.java
@@ -1,6 +1,6 @@
 package kr.ac.hongik.dsc2023.ydy.team1.core.controller;
 
-import kr.ac.hongik.dsc2023.ydy.team1.core.dto.request.SearchItemRequestDTO;
+import kr.ac.hongik.dsc2023.ydy.team1.core.dto.request.impl.SearchItemRequestDTO;
 import kr.ac.hongik.dsc2023.ydy.team1.core.dto.response.ResponseWrapper;
 import kr.ac.hongik.dsc2023.ydy.team1.core.dto.response.SearchItemResponseDTO;
 import org.springframework.http.ResponseEntity;
