+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/dto/request/impl/ItemCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.dto.request;
+package kr.ac.hongik.dsc2023.ydy.team1.core.dto.request.impl;
 
 /**
  * Item 생성 요청을 표현하는 클래스
