+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/dto/request/impl/SearchItemRequestDTO.java
@@ -1,4 +1,4 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.dto.request;
+package kr.ac.hongik.dsc2023.ydy.team1.core.dto.request.impl;
 
 /**
  * 클라이언트의 Item 검색 요청을 표현하는 추상 클래스
