+++ src/main/java/kr/ac/hongik/dsc2023/ydy/team1/core/dto/request/impl/SearchStrength.java
@@ -1,4 +1,4 @@
-package kr.ac.hongik.dsc2023.ydy.team1.core.dto.request;
+package kr.ac.hongik.dsc2023.ydy.team1.core.dto.request.impl;
 
 /**
  * 상품 검색시, 검색 키워드와 검색 결과물의 일치 정도를 결정할때 사용하는 클래스
